test: put test in right spot
